feat: 100% SSOT for FPSS events + historical volume/count i64 widening (#372)
Drive every FPSS streaming event shape and every tick volume/count field
from the existing `fpss_event_schema.toml` / `tick_schema.toml` SSOT
across Rust core, Rust FFI, Python, TypeScript, Go, C header, C++. No
more hand-written duplicates; schema change = compile-fatal drift.

FPSS SSOT (5 SDKs now schema-driven):
- Extend `build_support/fpss_events.rs` generator (+995 LoC):
  - emit `ffi/src/fpss_event_structs.rs` (Rust FFI `#[repr(C)]` kind enum,
    `TdxFpss*` structs, ZERO_* consts, `TdxFpssEvent`)
  - emit `ffi/src/fpss_event_converter.rs` (schema-driven
    `fpss_event_to_ffi` converter; control-variant kind table kept as
    documented generator constant — 0..=6, 8..=11, 12 unknown-event)
  - emit `sdks/go/fpss_event_structs.h.inc` (C typedefs mirror)
  - emit `sdks/go/fpss_event_structs.go` (idiomatic Go types +
    `FpssEventKind` / `FpssCtrl*` constants + tagged `FpssEvent`)
- Obliterate hand-written duplicates:
  - `ffi/src/lib.rs` -442: kind enum, `TdxFpss*` structs, ZERO_*
    consts, 270-line `fpss_event_to_ffi` converter
  - `sdks/go/ffi_bridge.h` -87: FPSS typedef block → `#include`
  - `sdks/go/fpss.go` -120: event kind/ctrl kind/structs → generator
- Python/TS: `buffered_event.rs` already schema-driven; `next_event`
  dispatcher now returns typed `#[pyclass]` / `#[napi(object)]` for
  every variant including `Simple` and `RawData` (no PyDict fallback
  anywhere, no `ts_return_type` union duplication)
- Breaking Go renames (no active users — approved):
  `FpssOpenInterestData` → `FpssOpenInterest`,
  `FpssControlData` → `FpssControl`, top-level `RawCode` / `RawPayload`
  → `RawData *FpssRawData`. Callers in `sdks/go/`, docs, and examples
  updated.

Issue #372 — historical volume/count i64 widening:
- New `eod_num64` pseudo-type in `tick_schema.toml` + generator.
  `EodTick.volume` / `EodTick.count`: `eod_num` → `eod_num64`.
  `OhlcTick.volume` / `OhlcTick.count`: `i32` → `i64`.
- Emit matching `eod_num64(row, idx) -> Result<i64, DecodeError>`
  helper in the generated decode prelude.
- Propagated across every language target:
  - Rust core (`crates/tdbe/src/types/tick.rs`): `i64`
  - Python pyclass: `i64` (pyo3 arbitrary-precision)
  - TS napi struct: `BigInt` (precision-preserving)
  - Go struct: `int64`
  - C/C++ struct: `int64_t`
  - Rust FFI tick layout: `i64` (128-byte layout still matches;
    FFI layout tests pass)
  - CLI: `raw_i64` helper replaces `raw_i32` for EOD/OHLC volume+count
    render call-sites

Validation:
- `cargo fmt --all -- --check`, `cargo clippy --workspace -- -D warnings`
  clean
- `cargo test --workspace`: 296 tests, 0 failures
- `go build ./...` + `go test ./...` clean (FFI layout assertions hold)
- `napi build --platform --release` clean; `index.d.ts` surfaces
  `bigint` for `EodTick.volume`/`count` + `OhlcTick.volume`/`count`;
  discriminated `nextEvent()` union has no duplicate arms
- `generate_sdk_surfaces --check` drift-free

Net LoC: +1293 −1525 across 27 files + 6 new generated files;
687 hand-written lines replaced by a 995-line generator.

+++ sdks/go/tick_structs.go
@@ -22,8 +22,8 @@ type EodTick struct {
 	High           float64  `json:"high"`
 	Low            float64  `json:"low"`
 	Close          float64  `json:"close"`
-	Volume         int      `json:"volume"`
-	Count          int      `json:"count"`
+	Volume         int64    `json:"volume"`
+	Count          int64    `json:"count"`
 	BidSize        int      `json:"bid_size"`
 	BidExchange    int      `json:"bid_exchange"`
 	Bid            float64  `json:"bid"`
@@ -106,8 +106,8 @@ type OhlcTick struct {
 	High           float64  `json:"high"`
 	Low            float64  `json:"low"`
 	Close          float64  `json:"close"`
-	Volume         int      `json:"volume"`
-	Count          int      `json:"count"`
+	Volume         int64    `json:"volume"`
+	Count          int64    `json:"count"`
 	Date           int      `json:"date"`
 	Expiration     int32    `json:"expiration,omitempty"`
 	Strike         float64  `json:"strike,omitempty"`
